BJS2-40466 moved back RequestStatus.java to enity package
