+++ src/main/java/school/faang/user_service/dto/mentorship/MentorshipRequestDto.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import school.faang.user_service.entity.mentorship.RequestStatus;
+import school.faang.user_service.entity.User.RequestStatus;
 
 @Data
 @Builder
