+++ src/main/java/school/faang/user_service/entity/goal/GoalInvitation.java
@@ -6,7 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-import school.faang.user_service.entity.mentorship.RequestStatus;
+import school.faang.user_service.entity.User.RequestStatus;
 import school.faang.user_service.entity.User;
 
 import java.time.LocalDateTime;
