+++ src/main/java/school/faang/user_service/entity/mentorship/MentorshipRequest.java
@@ -36,7 +36,7 @@ public class MentorshipRequest {
 
     @Column(name = "status", nullable = false)
     @Enumerated(EnumType.ORDINAL)
-    private RequestStatus status;
+    private User.RequestStatus status;
 
     @Column(name = "rejection_reason", length = 4096)
     private String rejectionReason;
