--- src/main/java/school/faang/user_service/entity/mentorship/RequestStatus.java
@@ -1,7 +1,0 @@
-package school.faang.user_service.entity.mentorship;
-
-public enum RequestStatus {
-    PENDING,
-    ACCEPTED,
-    REJECTED
-}
